+++ lib/ruby-next/language/proposed.rb
@@ -4,3 +4,6 @@
 
 require "ruby-next/language/rewriters/method_reference"
 RubyNext::Language.rewriters << RubyNext::Language::Rewriters::MethodReference
+
+require "ruby-next/language/rewriters/shorthand_hash"
+RubyNext::Language.rewriters << RubyNext::Language::Rewriters::ShorthandHash